rename typo to typography
