+++ src/js/view/elements/typography/TextBodySmall.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { TypoProps } from "./TypoProps";
+import { TypographyProps } from "./TypographyProps";
 
-export const TextBodySmall = (props: TypoProps) => {
+export const TextBodySmall = (props: TypographyProps) => {
   const { children, className = "" } = props;
 
   return (
