+++ src/js/view/elements/typography/TitleH2.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { TypoProps } from "./TypoProps";
+import { TypographyProps } from "./TypographyProps";
 
-export const TitleH2 = (props: TypoProps) => {
+export const TitleH2 = (props: TypographyProps) => {
   const { children, className = "" } = props;
 
   return <h2 className={`text-h2 font-bold ${className}`}>{children}</h2>;
