+++ src/js/view/elements/typography/TitleH3.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { TypoProps } from "./TypoProps";
+import { TypographyProps } from "./TypographyProps";
 
-export const TitleH3 = (props: TypoProps) => {
+export const TitleH3 = (props: TypographyProps) => {
   const { children, className = "" } = props;
 
   return <h3 className={`text-h3 font-bold ${className}`}>{children}</h3>;
