+++ src/styles/_variables.css
@@ -1,22 +1,1 @@
 
-:root{
-    /* Colors */
-    --primary: #8BFFC1;
-    --primary-hover: #57EBAC;
-    --primary-pressed: #39D996;
-    --light: #fff;
-    --dark-1: #262631;
-    --dark-2: #1C1C28;
-    --dark-3: #252732;
-
-    --bg-color: var(--dark-1);
-
-    /* Font colors */
-    --font-color-bright: var(--light);
-    --font-color-dark: var(--dark-2);
-    --font-color: #C7C9D9;
-
-    /* Radius */
-    --radius-primary: 12px;
-}
-
